Extract shared upgrade flow into UpgradeFlowService
