+++ apps/web/src/app/billing/individual/upgrade/services/index.ts
@@ -1,1 +1,2 @@
 export * from "./unified-upgrade-prompt.service";
+export * from "./upgrade-flow.service";
